practice-app/group5_api/requirements.txt
asgiref==3.3.4
psycopg2==2.8.6
certifi==2021.5.30
chardet==4.0.0
config==0.5.0.post0
decouple==0.0.7
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
Pygments==2.9.0
setuptools==41.2.0